+++ pipeline/json_test.go
@@ -237,7 +237,8 @@ func TestPipeline_MarshalJSON(t *testing.T) {
                     "endpoint": "",
                     "headers": null,
                     "captureResponse": false,
-                    "timeout": 0
+                    "timeout": 0,
+                    "skipSSLVerification": false
                 }
             ],
             "tcp": null,