+++ pipeline/tick/alert.go
@@ -69,7 +69,8 @@ func (n *AlertNode) Build(a *pipeline.AlertNode) (ast.Node, error) {
 		n.DotRemoveZeroValue("post", h.URL).
 			Dot("endpoint", h.Endpoint).
 			DotIf("captureResponse", h.CaptureResponseFlag).
-			Dot("timeout", h.Timeout)
+			Dot("timeout", h.Timeout).
+			DotIf("skipSSLVerification", h.SkipSSLVerificationFlag)
 
 		var headers []string
 		for k := range h.Headers {
@@ -120,7 +121,14 @@ func (n *AlertNode) Build(a *pipeline.AlertNode) (ast.Node, error) {
 
 	for _, h := range a.PagerDuty2Handlers {
 		n.Dot("pagerDuty2").
-			Dot("serviceKey", h.ServiceKey)
+			Dot("routingKey", h.RoutingKey)
+		for _, l := range h.Links {
+			if len(l.Text) > 0 {
+				n.Dot("link", l.Href, l.Text)
+			} else {
+				n.Dot("link", l.Href)
+			}
+		}
 	}
 
 	for _, h := range a.PushoverHandlers {
@@ -137,6 +145,16 @@ func (n *AlertNode) Build(a *pipeline.AlertNode) (ast.Node, error) {
 		n.Dot("sensu").
 			Dot("source", h.Source).
 			Dot("handlers", args(h.HandlersList)...)
+
+		// Use stable key order
+		keys := make([]string, 0, len(h.MetadataMap))
+		for k := range h.MetadataMap {
+			keys = append(keys, k)
+		}
+		sort.Strings(keys)
+		for _, k := range keys {
+			n.Dot("metadata", k, h.MetadataMap[k])
+		}
 	}
 
 	for _, h := range a.SlackHandlers {