+++ CHANGELOG.md
@@ -5,6 +5,41 @@
 ### Features
 
 - [#1965](https://github.com/influxdata/kapacitor/pull/1965): Alert handler for Microsoft Teams.
+## unreleased
+
+### Features
+
+- [#2154](https://github.com/influxdata/kapacitor/pull/2154): Add ability to skip ssl verification with an alert post node. Thanks @itsHabib!
+
+### Bugfixes
+
+- [#2167](https://github.com/influxdata/kapacitor/pull/2167): Use default transport consistently.
+- [#2144](https://github.com/influxdata/kapacitor/issues/2144): Fix deadlock in barrier node when delete is used.
+- [#2186](https://github.com/influxdata/kapacitor/pull/2186): Make RPM create files with correct ownership on install.
+- [#2189](https://github.com/influxdata/kapacitor/pull/2189): Delete group stats when a group is deleted
+
+## v1.5.2 [2018-12-12]
+
+### Features
+
+- [#2095](https://github.com/influxdata/kapacitor/issues/2095): Add barrier node support to join node.
+- [#1157](https://github.com/influxdata/kapacitor/issues/1157): Add ability to expire groups using the barrier node.
+- [#2099](https://github.com/influxdata/kapacitor/issues/2099): Add `alert/persist-topics` to config
+- [#2101](https://github.com/influxdata/kapacitor/issues/2101): Add multiple field support to the change detect node.
+- [#1961](https://github.com/influxdata/kapacitor/pull/1961): Add links to pagerduty2 alerts
+- [#1974](https://github.com/influxdata/kapacitor/issues/1974): Add additional metadata to Sensu alerts.
+
+### Bugfixes
+
+- [#2048](https://github.com/influxdata/kapacitor/pull/2048): Fix join not catching up fast enough after a pause in the data stream.
+
+## v1.5.1 [2018-08-06]
+
+### Bugfixes
+
+- [#1938](https://github.com/influxdata/kapacitor/issues/1938): pagerduty2 should use routingKey rather than serviceKey
+- [#1982](https://github.com/influxdata/kapacitor/pull/1982): Fix KafkaTopic not working from TICKscript
+- [#1989](https://github.com/influxdata/kapacitor/pull/1989): Improve Kafka alert throughput.
 
 ## v1.5.0 [2018-05-17]